refactor: rename and reposition files
