+++ app/components/DoubtsSection/AccordionDetails/style.module.css
