+++ app/components/DoubtsSection/AccordionSummary/style.module.css
