+++ app/components/DoubtsSection/index.tsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react'
 import { Tables } from '@/app/protocols'
 import { DoubtsContentType, DoubtsType } from './types'
 import { Tab, Tabs } from '@mui/material'
-import { FAQAccordionSummary } from './accordionSummary'
+import { FAQAccordionSummary } from './AccordionSummary'
 import style from './style.module.css'
 
 interface DoubtsProps {
